Ceci est la combinaison de 2 commits.
(PC-4996) moving: move HomeNavigator to Home feature
