+++ src/features/deeplinks/types.ts
@@ -1,4 +1,4 @@
-import { RouteParams, RootStackParamList } from 'features/navigation/RootNavigator'
+import { RouteParams, RootStackParamList } from 'features/home/navigation/HomeNavigator'
 
 export interface DeeplinkParts {
   routeName: string
